rename: getXXXXよりもAPIから取得するという意味でfetchを使った方がいいと思ったので変更
